+++ config/config.exs
@@ -1,4 +1,5 @@
 import Config
 
 config :elixir, :time_zone_database, Tzdata.TimeZoneDatabase
+config :gettext, :default_locale, "fr"
 config :phoenix, :json_library, Jason