+++ test/support/gettext.ex
@@ -1,0 +1,4 @@
+defmodule Doggo.Gettext do
+  @moduledoc false
+  use Gettext, otp_app: :doggo, priv: "test/support/gettext"
+end